simplified distribution

+++ src/main/java/eu/zavadil/openpublisher/api/ArticlesController.java
@@ -2,12 +2,14 @@ package eu.zavadil.openpublisher.api;
 
 import eu.zavadil.java.UrlBuilder;
 import eu.zavadil.java.spring.common.exceptions.BadRequestException;
+import eu.zavadil.java.spring.common.exceptions.NotAuthorizedException;
 import eu.zavadil.java.spring.common.exceptions.ResourceNotFoundException;
 import eu.zavadil.java.spring.common.exceptions.ServerErrorException;
 import eu.zavadil.java.spring.common.paging.JsonPage;
 import eu.zavadil.java.spring.common.paging.JsonPageImpl;
 import eu.zavadil.java.util.FileNameUtils;
 import eu.zavadil.openpublisher.data.article.Article;
+import eu.zavadil.openpublisher.data.article.ArticleState;
 import eu.zavadil.openpublisher.data.article.ArticleStub;
 import eu.zavadil.openpublisher.data.articleImage.ArticleImage;
 import eu.zavadil.openpublisher.data.user.User;
@@ -47,30 +49,92 @@ public class ArticlesController {
 		@RequestParam(defaultValue = "") String search,
 		@RequestParam(defaultValue = "") String sorting
 	) {
-		// todo: limit to owned articles for guests
-		return JsonPageImpl.of(this.articlesService.search(page, size, search, sorting));
+		return JsonPageImpl.of(
+			user.getUserRole().isAccessAllArticles() ? this.articlesService.search(page, size, search, sorting)
+				: this.articlesService.searchByUser(user.getId(), page, size, search, sorting)
+		);
+	}
+
+	@GetMapping("by-user/{id}")
+	@Secured({UserRole.ADMIN_ROLE_NAME})
+	public JsonPage<Article> loadPagedByUser(
+		@PathVariable int id,
+		@RequestParam(defaultValue = "0") int page,
+		@RequestParam(defaultValue = "10") int size,
+		@RequestParam(defaultValue = "") String search,
+		@RequestParam(defaultValue = "") String sorting
+	) {
+		return JsonPageImpl.of(this.articlesService.searchByUser(id, page, size, search, sorting));
+	}
+
+	@GetMapping("by-destination/{id}")
+	@Secured({UserRole.ADMIN_ROLE_NAME})
+	public JsonPage<Article> loadPagedByDestination(
+		@PathVariable int id,
+		@RequestParam(defaultValue = "0") int page,
+		@RequestParam(defaultValue = "10") int size,
+		@RequestParam(defaultValue = "") String search,
+		@RequestParam(defaultValue = "") String sorting
+	) {
+		return JsonPageImpl.of(this.articlesService.searchByDestination(id, page, size, search, sorting));
 	}
 
 	@GetMapping("{id}")
-	public ArticleStub load(@PathVariable int id) {
-		return this.articlesService.loadById(id);
+	public ArticleStub load(
+		@AuthenticationPrincipal User user,
+		@PathVariable int id
+	) {
+		ArticleStub article = this.articlesService.loadById(id);
+		if (article == null) throw new ResourceNotFoundException("Article with id " + id + " not found");
+		if (!this.articlesService.canAccess(user, article))
+			throw new NotAuthorizedException("K tomuto článku nemáte právo přístupu");
+		return article;
 	}
 
 	@PostMapping("")
-	@Secured({UserRole.EDITOR_ROLE_NAME, UserRole.ADMIN_ROLE_NAME})
+	@Secured({UserRole.EDITOR_ROLE_NAME})
 	public ArticleStub insert(@RequestBody ArticleStub document) {
 		document.setId(null);
 		return this.articlesService.save(document);
 	}
 
 	@PutMapping("{id}")
-	public ArticleStub update(@PathVariable int id, @RequestBody ArticleStub document) {
-		document.setId(id);
-		return this.articlesService.save(document);
+	public ArticleStub update(
+		@AuthenticationPrincipal User user,
+		@PathVariable int id,
+		@RequestBody ArticleStub document
+	) {
+		ArticleStub existing = this.articlesService.loadById(id);
+		if (existing == null) throw new ResourceNotFoundException("Článek", id);
+		if (!this.articlesService.canAccess(user, existing))
+			throw new NotAuthorizedException("K tomuto článku nemáte právo přístupu");
+		if (user.getUserRole() == UserRole.Guest && existing.getArticleState() == ArticleState.Approved) {
+			throw new NotAuthorizedException("Jako externí partner nemáte možnost upravovat schválené články");
+		}
+		if (user.getUserRole() == UserRole.Guest && document.getArticleState() == ArticleState.Approved) {
+			throw new NotAuthorizedException("Jako externí partner nemáte možnost schvalovat články");
+		}
+
+		// public fields
+		existing.setArticleState(document.getArticleState());
+		existing.setImageName(document.getImageName());
+		existing.setHeader(document.getHeader());
+		existing.setPreviewText(document.getPreviewText());
+		existing.setContentHtml(document.getContentHtml());
+		existing.setPublishDate(document.getPublishDate());
+
+		// admin fields
+		if (user.getUserRole().isAccessAllArticles()) {
+			existing.setPartnerId(document.getPartnerId());
+			existing.setOwnerId(document.getOwnerId());
+			existing.setDestinationId(document.getDestinationId());
+		}
+
+		return this.articlesService.save(existing);
 	}
 
 	@DeleteMapping("{id}")
-	@Secured({UserRole.EDITOR_ROLE_NAME, UserRole.ADMIN_ROLE_NAME})
+	@Secured({UserRole.ADMIN_ROLE_NAME})
 	public void delete(@PathVariable int id) {
 		this.articlesService.delete(id);
 	}
@@ -86,9 +150,13 @@ public class ArticlesController {
 
 	@PostMapping("{id}/images")
 	public ArticleImage uploadImage(
+		@AuthenticationPrincipal User user,
 		@PathVariable int id,
 		@RequestParam("image") MultipartFile file
 	) {
+		if (!this.articlesService.canAccess(user, id))
+			throw new NotAuthorizedException("K tomuto článku nemáte právo přístupu");
+
 		String originalFileName = FileNameUtils.extractFileName(file.getOriginalFilename());
 		try {
 			return this.articlesService.uploadArticleImage(id, originalFileName, file.getBytes());
@@ -100,9 +168,13 @@ public class ArticlesController {
 
 	@DeleteMapping("{id}/images/{imageName}")
 	public void deleteImage(
+		@AuthenticationPrincipal User user,
 		@PathVariable int id,
 		@PathVariable String imageName
 	) {
+		if (!this.articlesService.canAccess(user, id))
+			throw new NotAuthorizedException("K tomuto článku nemáte právo přístupu");
+
 		this.articlesService.deleteImage(id, imageName);
 	}
 
@@ -135,28 +207,31 @@ public class ArticlesController {
 	private AccessService accessService;
 
 	@PostMapping("{id}/grant-guest-access/{partnerEmail}")
-	@Secured({UserRole.EDITOR_ROLE_NAME, UserRole.ADMIN_ROLE_NAME})
+	@Secured({UserRole.EDITOR_ROLE_NAME})
 	public String grantGuestAccess(
+		@AuthenticationPrincipal User user,
 		@PathVariable int id,
 		@PathVariable String partnerEmail
 	) {
 		ArticleStub article = this.articlesService.loadById(id);
 		if (article == null) throw new ResourceNotFoundException("Article not found!");
+		if (!this.articlesService.canAccess(user, article))
+			throw new NotAuthorizedException("K tomuto článku nemáte právo přístupu");
 
-		User user = this.usersService.loadByEmail(partnerEmail);
-		if (user == null) {
-			user = new User();
-			user.setEmail(partnerEmail);
-			user.setUserRole(UserRole.Guest);
+		User partner = this.usersService.loadByEmail(partnerEmail);
+		if (partner == null) {
+			partner = new User();
+			partner.setEmail(partnerEmail);
+			partner.setUserRole(UserRole.Guest);
 		}
 
-		user.setActive(true);
-		this.usersService.save(user);
+		partner.setActive(true);
+		this.usersService.save(partner);
 
-		article.setPartnerId(user.getId());
+		article.setPartnerId(partner.getId());
 		this.articlesService.save(article);
 
-		String token = this.accessService.createEncodedAccessToken(user);
+		String token = this.accessService.createEncodedAccessToken(partner);
 		String url = UrlBuilder.of(this.urlBase)
 			.addPath("clanky/detail")
 			.addPath(article.getId().toString())
